+++ src/main/java/no2/worldthreader/WorldThreaderMod.java
@@ -1,7 +1,9 @@
 package no2.worldthreader;
 
-import no2.worldthreader.init.ModGameRules;
 import net.fabricmc.api.ModInitializer;
+import net.minecraft.world.level.block.Blocks;
+import no2.worldthreader.common.mixin_support.interfaces.BeforeThreadingInitialization;
+import no2.worldthreader.init.ModGameRules;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -13,5 +15,11 @@ public class WorldThreaderMod implements ModInitializer {
 	@Override
 	public void onInitialize() {
 		ModGameRules.registerGameRules();
+	}
+
+	public static void initializeBeforeThreading() {
+		((BeforeThreadingInitialization) Blocks.CARVED_PUMPKIN).worldthreader$initBeforeThreading();
+		((BeforeThreadingInitialization) Blocks.END_PORTAL_FRAME).worldthreader$initBeforeThreading();
+		((BeforeThreadingInitialization) Blocks.WITHER_SKELETON_SKULL).worldthreader$initBeforeThreading();
 	}
 }