+++ src/main/java/no2/worldthreader/common/mixin_support/interfaces/BeforeThreadingInitialization.java
@@ -1,0 +1,5 @@
+package no2.worldthreader.common.mixin_support.interfaces;
+
+public interface BeforeThreadingInitialization {
+    void worldthreader$initBeforeThreading();
+}