+++ src/main/java/no2/worldthreader/common/thread/WorldThreadingManager.java
@@ -51,6 +51,8 @@ public class WorldThreadingManager {
 
 
 	public WorldThreadingManager(MinecraftServer server) {
+		WorldThreaderMod.initializeBeforeThreading();
+
 		this.server = server;
 		this.tickBarrier = new Phaser();
 		this.withinTickBarrier = new Phaser();